move CreateObservablePart
