+++ src/core/observable-api/unique-behavior-subject.ts
@@ -1,4 +1,4 @@
-import { BehaviorSubject } from "rxjs";
+import { BehaviorSubject, distinctUntilChanged, map, Observable, shareReplay } from "rxjs";
 
 
 function deepFreeze<T>(inObj: T): T {
@@ -19,6 +19,30 @@ function deepFreeze<T>(inObj: T): T {
 
 export function naiveObjectComparison(objOne: any, objTwo: any): boolean {
     return JSON.stringify(objOne) === JSON.stringify(objTwo);
+}
+
+
+
+
+type MappingFunction<T, R> = (mappable: T) => R;
+type MemoizationFunction<R> = (previousResult: R, currentResult: R) => boolean;
+
+function defaultMemoization(previousValue: any, currentValue: any): boolean {
+  if (typeof previousValue === 'object' && typeof currentValue === 'object') {
+    return naiveObjectComparison(previousValue, currentValue);
+  }
+  return previousValue === currentValue;
+}
+export function CreateObservablePart<T, R> (
+	source$: Observable<T>,
+	mappingFunction: MappingFunction<T, R>,
+	memoizationFunction?: MemoizationFunction<R>
+): Observable<R> {
+	return source$.pipe(
+	  map(mappingFunction),
+	  distinctUntilChanged(memoizationFunction || defaultMemoization),
+	  shareReplay(1)
+	)
 }
